requirements.txt
Django==5.0
djangorestframework==3.14.0
funasr==0.8.0
redis==5.0.1
ffmpeg-python==0.2.0
numpy==1.26.3
python-dotenv==1.1.0
antlr4-python3-runtime==4.9.3
kaldiio==2.18.1
editdistance>=0.5.2
hydra-core>=1.3.2
jaconv
jamo
jieba
librosa
oss2
scipy>=1.4.1
sentencepiece
tensorboardX
umap-learn
